@@ -1,6 +1,6 @@
 {
   "name": "beantown",
-  "version": "0.1.51",
+  "version": "0.1.54",
   "private": true,
   "type": "module",
   "engines": {
@@ -40,6 +40,7 @@
     "@fortawesome/free-brands-svg-icons": "^6.1.1",
     "@fortawesome/free-solid-svg-icons": "^6.3.0",
     "@fortawesome/react-fontawesome": "^0.2.0",
+    "@jalgraves/react-components-library": "^1.0.74",
     "@types/node": "^17.0.23",
     "animejs": "^3.2.1",
     "babel-cli": "^6.26.0",
@@ -50,15 +51,15 @@
     "core-js": "^3.18.3",
     "eslint": "^8.0.1",
     "eslint-plugin-react-hooks": "^4.3.0",
-    "react": "^16.14.0",
-    "react-dom": "^16.14.0",
-    "react-hook-form": "^5.1.1",
-    "react-loader-spinner": "^5.3.4",
-    "react-router": "^6.15.0",
-    "react-router-dom": "^6.15.0",
-    "react-slick": "^0.29.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-hook-form": "^7.52.2",
+    "react-loader-spinner": "^6.1.6",
+    "react-router": "^6.26.0",
+    "react-router-dom": "^6.26.0",
+    "react-slick": "^0.30.2",
     "react-slideshow-image": "^4.2.1",
-    "styled-components": "^5.0.0",
+    "styled-components": "^6.0.8",
     "universal-cookie": "^4.0.3",
     "webpack": "^5.81.0",
     "webpack-cli": "^5.0.2"
