@@ -1,7 +1,8 @@
 import React from 'react'
-import { InfoSection } from './common.js'
 import { config } from '../../utils/main.js'
 import { sendEvent } from '../kafka/index.js'
+import { ContentContainer } from './container.js'
+import { GrubHubLink } from './grubHub/index.js'
 const COLORS = config.colors
 
 
@@ -14,15 +15,17 @@ const FAQ = () => {
       <h4>Do you take reservations?</h4>
       <p>No, but we do offer private parties for larger gatherings.</p>
       <h4>What time do you serve food 'til?</h4>
-      <p>12:00 am, though subject to change based on demand.</p>
+      <p>11:30, though subject to change based on demand.</p>
       <h4>Is the bar 21 plus?</h4>
-      <p>Before 6:00 pm all ages are welcome, after 6:00 21 plus only.</p>
+      <p>No persons under 21 are allowed to sit at the bar at anytime but before 6:00 pm all ages are welcome to sit at any table in the restaurant. After 6:00 minors must be accompanied by an adult.</p>
+      <p>You must be 21+ to play pool or be in the pool area</p>
       <h4>How many t.v.s do you have?</h4>
       <p>A lot! There are two large projection screens in the main pool area and many big screens scattered throughout the rest of the bar.</p>
       <h4>Can you put my game on?</h4>
-      <p>Most likely, if it's available.</p>
+      <p>Most likely, if it's available. Please keep in mind that we can only show three different at any given time.</p>
       <h4>Do you have takeout?</h4>
       <p>Yes</p>
+      <GrubHubLink/>
       <h4>How much does pool cost?</h4>
       <p>Pool tables are coin operated and are &#36;2.00 a game, there are change machines located in the pool area. Pool tables are first come first serve.</p>
       <h4>Can I have a job?</h4>
@@ -36,8 +39,8 @@ export const AboutInfo = () => {
     sendEvent(event)
   })
   return (
-    <div id="aboutInfo">
-      <InfoSection bgColor={COLORS.yellow} marginTop='3rem' slug="about-info">
+    <ContentContainer>
+      <ContentContainer margin="5% auto">
         <section>
           <h1>See Yah Laatah&trade;!</h1>
           <article id="aboutBeantownArticle">
@@ -46,10 +49,16 @@ export const AboutInfo = () => {
             to meet friends, or somewhere to feel at home while you’re traveling - Beantown Pub is the place to be!
           </article>
         </section>
-      </InfoSection>
-      <InfoSection bgColor={COLORS.darkGray} fontColor={COLORS.yellow} slug="faq">
+      </ContentContainer>
+      <ContentContainer
+        backgroundColor="beige"
+        borderRadius="5px"
+        h4Color={COLORS.black}
+        h4FontSize="1.5rem"
+        h4FontWeight="900"
+      >
         <FAQ/>
-      </InfoSection>
-    </div>
+      </ContentContainer>
+    </ContentContainer>
   )
 }
