@@ -4,7 +4,7 @@ const COLORS = config.colors
 const FONTS = config.fonts
 
 export const StyledContentContainer = styled.div`
-  background-color: ${props => props.backgroundColor || "beige"};
+  background-color: ${props => props.$backgroundColor || "beige"};
   color: ${props => props.fontColor || "unset"};
   display: flex;
   flex-flow: column wrap;
@@ -21,46 +21,6 @@ export const StyledContentContainer = styled.div`
     line-height: 170%;
     margin: 2rem auto 2rem auto;
   }
-  @media (min-width: 320px)
-  and (max-width: 768px)
-  and (orientation: portrait)
-  and (-webkit-min-device-pixel-ratio: 2) {
-    h1 {
-      font-size: 2rem;
-    }
-  }
-`
-
-export const StyledPartiesContainer = styled.div`
-  background-color: ${props => props.backgroundColor || "beige"};
-  color: ${props => props.fontColor || "unset"};
-  display: flex;
-  flex-flow: column wrap;
-  font-family: ${props => props.fontFamily || FONTS.content};
-  margin: ${props => props.margin || "5rem auto 2rem auto"};
-  max-width: 99vw;
-  padding: ${props => props.padding || "2rem"};
-  text-transform: ${props => props.textTransform || "unset"};
-  width: ${props => props.width || "100%"};
-  article {
-    font-size: 150%;
-    font-weight: 600;
-    line-height: 150%;
-    max-width: 80%;
-    margin: auto;
-    padding: .25rem;
-  }
-  h1 {
-    font-family: ${props => props.h1FontFamily || FONTS.button};
-    font-size: ${props => props.h1FontSize || "3rem"};
-    font-weight: ${props => props.h1FontWeight || "900"};
-    letter-spacing: .25rem;
-    line-height: ${props => props.h1LineHeight || "170%"};
-    margin: ${props => props.h1Margin || "3rem auto 2rem auto"};
-    padding: ${props => props.h1Padding || ".5rem 0"};
-    text-transform: uppercase;
-  }
-  
   @media (min-width: 320px)
   and (max-width: 768px)
   and (orientation: portrait)
@@ -88,7 +48,7 @@ export const StyledInfo = styled.div`
   font-family: ${FONTS.button};
   letter-spacing: .12rem;
   text-transform: uppercase;
-  color: ${props => props.fontColor || COLORS.black};
+  color: ${props => props.$fontColor || COLORS.black};
   line-height: 170%;
   h1 {
     padding: .5rem 0;
