@@ -1,6 +1,6 @@
 import React, { useState } from 'react'
 import { useForm } from 'react-hook-form'
-import { Icon } from '../../../content/icons/index.js'
+import { Icon } from "@jalgraves/react-components-library"
 import { SubmitButton, ToggleButton } from '../../../elements/buttons/main.js'
 import { StyledCheckoutForm } from './styles.js'
 import { PaymentForm } from '../../square/forms/main.js'
