@@ -14,8 +14,8 @@ const config = {
             "name": "index",
             "icon": "faBeer",
             "text": "Home",
-            "menu_list": true,
-            "top_menu": false,
+            "menuList": true,
+            "topMenu": false,
             "metadata": {
                 "title": "Beantown Pub",
                 "description": "The only pub in the world where you can drink a cold Sam Adams while viewing a cold Sam Adams",
@@ -27,8 +27,8 @@ const config = {
             "name": "about",
             "icon": "faInfo",
             "text": "About",
-            "menu_list": true,
-            "top_menu": true,
+            "menuList": true,
+            "topMenu": true,
             "metadata": {
                 "title": "About Beantown",
                 "description": "The only pub in the world where you can drink a cold Sam Adams while viewing a cold Sam Adams",
@@ -40,8 +40,8 @@ const config = {
             "name": "contact",
             "icon": "faEnvelope",
             "text": "Contact",
-            "menu_list": true,
-            "top_menu": true,
+            "menuList": true,
+            "topMenu": true,
             "metadata": {
                 "title": "Beantown Pub",
                 "description": "Contact information for Beantown Pub",
@@ -53,8 +53,8 @@ const config = {
             "name": "menu",
             "icon": "faUtensils",
             "text": "Menu",
-            "menu_list": true,
-            "top_menu": true,
+            "menuList": true,
+            "topMenu": true,
             "metadata": {
                 "title": "Beantown Pub",
                 "description": "Pub dining menu for Beantown Pub",
@@ -66,8 +66,8 @@ const config = {
             "name": "parties",
             "icon": "faCalendar",
             "text": "Private Parties",
-            "menu_list": true,
-            "top_menu": true,
+            "menuList": true,
+            "topMenu": true,
             "metadata": {
                 "title": "Private Events",
                 "description": "Private parties at Beantown Pub",
@@ -79,8 +79,8 @@ const config = {
             "name": "/merch/items",
             "icon": "faTshirt",
             "text": "Merch",
-            "menu_list": false,
-            "top_menu": false,
+            "menuList": false,
+            "topMenu": false,
             "metadata": {
                 "title": "Beantown Merch",
                 "description": "Beantown Pub merchandise for online purchase",
@@ -92,8 +92,8 @@ const config = {
             "name": "https://www.facebook.com/BeantownPubMA/",
             "icon": "faCamera",
             "text": "Photos",
-            "menu_list": true,
-            "top_menu": false,
+            "menuList": true,
+            "topMenu": false,
             "metadata": {
                 "title": "Beantown Pub",
                 "description": "",
@@ -104,8 +104,8 @@ const config = {
             "name": "/returns",
             "icon": "faTshirt",
             "text": "Returns",
-            "menu_list": false,
-            "top_menu": false,
+            "menuList": false,
+            "topMenu": false,
             "metadata": {
                 "title": "Return and Refund Policy",
                 "description": "Beantownpub.com's return and refund policy",
@@ -117,8 +117,8 @@ const config = {
             "name": "/privacy",
             "icon": "faLock",
             "text": "Privacy",
-            "menu_list": false,
-            "top_menu": false,
+            "menuList": false,
+            "topMenu": false,
             "metadata": {
                 "title": "Privacy Policy",
                 "description": "Beantownpub.com's privacy policy",
@@ -130,8 +130,8 @@ const config = {
             "name": "https://www.facebook.com/BeantownPubMA/",
             "icon": "faFacebook",
             "text": "Facebook",
-            "menu_list": true,
-            "top_menu": false,
+            "menuList": true,
+            "topMenu": false,
             "metadata": {
                 "title": "Beantown Pub",
                 "description": "",
@@ -142,8 +142,8 @@ const config = {
             "name": "https://www.instagram.com/explore/tags/beantownpub/?hl=en",
             "icon": "faInstagram",
             "text": "Instagram",
-            "menu_list": true,
-            "top_menu": false,
+            "menuList": true,
+            "topMenu": false,
             "metadata": {
                 "title": "Beantown Pub",
                 "description": "",
@@ -154,8 +154,8 @@ const config = {
             "name": "https://twitter.com/BeantownPub",
             "icon": "faTwitter",
             "text": "Twitter",
-            "menu_list": true,
-            "top_menu": false,
+            "menuList": true,
+            "topMenu": false,
             "metadata": {
                 "title": "Beantown Pub",
                 "description": "",
@@ -166,8 +166,8 @@ const config = {
             "name": "/error",
             "icon": "faBeer",
             "text": "Error",
-            "menu_list": false,
-            "top_menu": false,
+            "menuList": false,
+            "topMenu": false,
             "metadata": {
                 "title": "Beantown Error",
                 "description": "Beantownpub.com's error page",
@@ -182,6 +182,7 @@ const config = {
         "borderGray": "#dbd9d9",
         "borderShadow": "#383838",
         "boxShadowGray": "rgb(126 137 140 / 20%)",
+        "dodgerBlue": "#1e90ff",
         "gray": "#494040",
         "red": "#ed1c24",
         "navBarBlue": "rgba(5,46,102,0.90)",
@@ -193,6 +194,7 @@ const config = {
         "shadowGray": "rgb(147 147 147 / 25%)",
         "spinnerBlue": "#00BFFF",
         "dodgerBlue": "#1e90ff",
+        "grubHubOrange": "#e67e22",
         "cyan": "#3fc5f0",
         "black": "#000000",
         "topMenuBackground": "#494040",
